chore: fixes from CR
