+++ packages/core/nuxt-theme-module/theme/layouts/account.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import AppHeader from '~/components/AppHeader.vue';
+import AppHeader from '~/components/header/AppHeader.vue';
 import BottomNavigation from '~/components/BottomNavigation.vue';
 import TopBar from '~/components/TopBar.vue';
 import CartSidebar from '~/components/CartSidebar.vue';
